+++ example/gui/gif_player/gif_player_demo.c
@@ -1,4 +1,8 @@
 #include "bda_controls.h"
+#include "bda_input.h"
+#include "bda_memory.h"
+#include "bda_time.h"
+#include "bda_window.h"
 
 #define GIF_CONTROL_ID 0x241u